Modify TxInsReference to incorporate the possibility that the reference input may not be witnessing a particular tx input. This is represented with the Maybe TxIn type.
We wrap our reference input with a newtype to
make it clear which tx input is the reference input.

We modify the BabbageEra's makeShelleyTransacionBody case. We must
include the inputs that are potentially being witnessed by a reference
input (via a script). We also include reference input related datums and
redeemrs, again if they are actually witnessing a tx input.

We propagate the user specified script language when using a reference
script in order to calculate the correct integrity hash.